Move SingleLetterAbbrivation into languages common
